requirements.txt
Django==3.1.4
djangorestframework==3.12.2
gunicorn==20.0.4
psycopg2==2.8.6
maya
requests
BeautifulSoup4
django-rest-auth
elasticsearch
django-bulk-update-or-create
selenium
channels